Fixed maranget's algorithm bug
The issue was to do with references. In the pattern match matrix, I was
appending rows to new matrices (as lists), but forgetting to clone the
objects first. This meant that it was possible for two rows in two
separate matrices to be coupled. Still not entirely happy with how this
is implemented, would love it if there is a better way?