ADTs can now reference function types
I.e. you can now make a list of functions. Also, fixed a bug in the
python code generator. I failed to account for the fact that there
can be two lambdas in the same 'scope' and that they need to be named
uniquely. This was giving rise to some very weird behaviours.